Clean up narrowing warnings
This change is moved out from #118 PR.

There are a few warnings about integer narrowing in the codebase. These
warning produce a lot of output on Windows and make it easy to miss
actually important warnings from new changes.

This change added explicit static casts for integers narrowing.

Testing: no functional changes, no new tests.

+++ src/parameter_descriptor.cpp
@@ -338,12 +338,13 @@ SQLRETURN ParameterDescriptor::SetValue(idx_t rec_idx) {
 			return SQL_ERROR;
 		}
 		if (precision <= Decimal::MAX_WIDTH_INT64) {
-			value = Value::DECIMAL(Load<int64_t>(dataptr), precision, scale);
+			value =
+			    Value::DECIMAL(Load<int64_t>(dataptr), static_cast<uint8_t>(precision), static_cast<uint8_t>(scale));
 		} else {
 			hugeint_t dec_value;
 			memcpy(&dec_value.lower, dataptr, sizeof(dec_value.lower));
 			memcpy(&dec_value.upper, dataptr + sizeof(dec_value.lower), sizeof(dec_value.upper));
-			value = Value::DECIMAL(dec_value, precision, scale);
+			value = Value::DECIMAL(dec_value, static_cast<uint8_t>(precision), static_cast<uint8_t>(scale));
 		}
 		break;
 	}
